Fix /api/admin/hubbell/suggestions 500: so_id varchar/int join
The suggestions list endpoint and the AI review endpoint both joined
agility_so_header.so_id (varchar) directly to
hubbell_document_suggestions.so_id (integer), which Postgres rejects at
plan time with "operator does not exist: character varying = integer" —
identical bug to the one already fixed on /api/hubbell/checks. Codex
caught this end-to-end from the agent-side CLI.

Fix: cast soh.so_id::int on both joins (the varchar may carry leading
zeros, so int cast normalizes both sides safely). Same pattern used by
other routes that touch agility_so_header.

Also wrapped the suggestions GET in try/catch so future 500s surface the
actual error message in the response body. Matches what
/api/hubbell/checks does. Helps CLI/agent callers diagnose without
poking Vercel logs.

Unblocks scripts/hubbell-review/index.ts (`pull` was 500ing on the
suggestions GET).

+++ app/api/admin/hubbell/documents/ai-review/route.ts
@@ -204,7 +204,7 @@ export async function POST(req: NextRequest) {
     FROM bids.hubbell_document_suggestions s
     JOIN bids.hubbell_documents d ON d.id = s.document_id
     LEFT JOIN public.agility_so_header soh
-      ON soh.so_id = s.so_id AND soh.is_deleted = false
+      ON soh.so_id::int = s.so_id AND soh.is_deleted = false
     WHERE s.status = 'pending'
     ${docFilter}
     ORDER BY s.document_id, s.confidence DESC, s.suggested_at DESC
